migrate SummaryBuilderSpec
