--- test-tests/shared/src/main/scala/zio/test/ReportingTestUtils.scala
@@ -1,29 +1,0 @@
-package zio.test
-
-import scala.{ Console => SConsole }
-
-object ReportingTestUtils {
-  def expectedSuccess(label: String): String =
-    green("+") + " " + label + "\n"
-
-  def expectedFailure(label: String): String =
-    red("- " + label) + "\n"
-
-  def withOffset(n: Int)(s: String): String =
-    " " * n + s
-
-  def green(s: String): String =
-    SConsole.GREEN + s + SConsole.RESET
-
-  def red(s: String): String =
-    SConsole.RED + s + SConsole.RESET
-
-  def blue(s: String): String =
-    SConsole.BLUE + s + SConsole.RESET
-
-  def cyan(s: String): String =
-    SConsole.CYAN + s + SConsole.RESET
-
-  def yellowThenCyan(s: String): String =
-    SConsole.YELLOW + s + SConsole.CYAN
-}
